requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.7
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
fastapi==0.103.2
h11==0.14.0
idna==3.4
Markdown==3.5
Pillow==10.0.1
pydantic==2.4.2
pydantic_core==2.10.1
python-multipart==0.0.5
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
